Reverted to using a simple fix for loading recipes from storage at start.
If the file path exists eg. xxxx/recipe

then do another check if it isn't a directory, throw an error prompting user to change it to allow the saving of recipes to work properly.

This is to fix bug#1 found in our discord (you may request access for this through our issues or email us)